// shared/node/v0.2.0/types.d.ts
export type NodeContext = {
    callbacksAfterUpdate: any
    componentModels: { [x: string]: any }
    connectionPulsingCallbackScheduled: boolean
    connectionsToPulse: any
    connectionsToPulseChanged: boolean
    currentFrameTime: number
    debugInspectors: any
    debugInspectorsEnabled: boolean
    editorConnection: any
    eventEmitter: { _events: any, _eventsCount: number, _maxListeners: number }
    eventSenderEmitter: { _events: any, _eventsCount: number, _maxListeners: number }
    frameNumber: number
    globalValues: any
    globalsEventEmitter: { _events: any, _eventsCount: number, _maxListeners: number }
    graphModel: GraphModel
    isUpdating: boolean
    nodeRegister: { _constructors: any, context: any }
    onClosePopup: () => any
    onShowPopup: () => any
    platform: { webSocketOptions: any, requestUpdate: any, getCurrentTime: any, objectToString: any }
    rootComponent: { id: string, context: any, _dirty: boolean, _inputs: any, _inputValues: any }
    runningInEditor: boolean
    setNodeFocused: () => any
    styles: { graphModel: any, nodeRegister: any, styles: any, getNodeScope: any }
    timerScheduler: { runningTimers: any, newTimers: any, requestFrame: any }
    updateIteration: number
    variants: any
    warningTypes: any
    _dirtyNodes: any
    _outputHistory: any
    _signalHistory: any
}

export type GraphModel = {
    componentIndex: any
    componentToBundleMap: { size: number }
    components: { [x: string]: any }
    listeners: { [x: string]: any }
    listenersWithRefs: { [x: string]: any }
    metadata: { styles: { colors: { [x: string]: any }, text: { [x: string]: any } } }
    rootComponent: string
    routerIndex: { routers: any, pages: any }
    settings: { navigationPathType: 'path' }
    variants: any
    on(listener: string, callback: (node: GraphModelNode) => void): void
}

export type GraphModelNode = {
    children: GraphModelNode[]
    component: any
    id: string
    inputPorts: any
    inputs: any
    listeners: { parameterUpdated: any }
    listenersWithRefs: { parameterUpdated: any, variantUpdated: any, inputPortRemoved: any, outputPortRemoved: any }
    outputPorts: any
    outputs: any
    parameters: { version: 'v0.1.0', [x: string]: any }
    parent: GraphModelNode
    type: string
    on(listener: string, callback: ({ name, state, value }) => void): void
}